--- apps/events/models.py
@@ -1,243 +1,0 @@
-
-"""
-    Handles all models related to events :
-        - Event : Information about an event in Fest
-        - Tab : Information about an event tab in Fest
-        
-        This models.py has been combined with apps/portals/models.py
-
-"""
-# Django
-from django.db import models
-from django.contrib.auth.models import User
-from django.conf import settings
-from django.utils import timezone
-# Apps
-import select2.fields
-# Decorators
-# Models
-from apps.walls.models import Wall, Post
-from apps.users.models import User,UserProfile,Team,ERPProfile
-# Forms
-# View functions
-# Misc
-from misc.utils import *
-# Python
-from configs.settings import FEST_NAME
-import select2.models
-import select2.forms
-from django.core.management import call_command
-import os
-from django.db.models.signals import post_save, post_delete
-from django.dispatch import receiver
-
-EVENT_VENUES=settings.EVENT_VENUES
-
-if FEST_NAME=='Saarang':
-	EVENT_CATEGORIES = settings.EVENT_CATEGORIES
-else:
-	EVENT_CATEGORIES = (
-		('Aerofest', 'Aerofest'),
-		('Coding', 'Coding'),
-		('Design and build', 'Design and build'),
-		('Involve', 'Involve'),
-		('Quizzes', 'Quizzes'),
-		('Online', 'Online'),
-		('Department Flagship Event', 'Department Flagship Event'),
-		('Spotlight', 'Spotlight'),
-		('Workshops', 'Workshops'),
-		('Exhibitions', 'Exhibitions and Shows'),
-		('Miscellaneous', 'Miscellaneous'),
-		('Sampark', 'Sampark'),
-		('B- Events','B- Events'),
-		('Associated Events','Associated Events'),
-	)
-
-EVENT_TYPE = (
-    ('online', 'Online Registration'),
-    ('onspot', 'On-spot Registration'),
-    ('noreg','No Registration'),
-)
-    
-class Event(models.Model):
-    """
-        An Event model which defines every event happening in the fest.
-        #question : Should FAQ be a different model ? Why was it kept like this in saarang ?
-    """
-    
-    # Basic information
-    name                = models.CharField(max_length=50, unique=True)
-    short_description   = models.CharField(max_length=250, blank=True)
-    event_type          = models.CharField(max_length=100, choices=EVENT_TYPE, blank=True, null=True)
-    category            = models.CharField(max_length=100, choices=EVENT_CATEGORIES)
-    
-    # Event - Team specific information
-    has_tdp             = models.BooleanField(default=False)
-    team_size_min       = models.IntegerField(default=1, blank=True, null=True) # Team event or not can be got using these fields
-    team_size_max       = models.IntegerField(default=1, blank=True, null=True)
-    
-    # Event registration information
-    registration_starts = models.DateTimeField(blank=True, null=True) # Using the two dates it can be found if registration is on or not
-    registration_ends   = models.DateTimeField(blank=True, null=True)
-    
-    # Email ids specific to the Event : google_group and the corresponding shaatsra_email_id
-    google_group        = models.EmailField(max_length=100, blank=True, null=True)
-    email               = models.EmailField(max_length=100, blank=True, null=True)
-
-	# List of registered participants
-    users_registered = models.ManyToManyField(User, blank=True, null=True,related_name='events_registered')
-    teams_registered = models.ManyToManyField(Team, blank=True, null=True,related_name='events_registered')
-    #added by Akshay/Arun
-    coords = models.ManyToManyField(ERPProfile, null=True, blank=True, related_name='coord_events')
-    long_description=models.TextField(blank = True, null=True)
-    google_form=models.URLField(blank=True, null=True)
-    event_image=models.ImageField(blank=True, null=True, upload_to='events')
-    extra_info = models.BooleanField(blank=True, default=False)
-
-    # Extra mainsite information
-    is_visible = models.BooleanField(default=True) # On the mainsite
-    
-    # Some properties to make some conditions easier
-    @property
-    def is_team_event(self):        
-        return self.team_size_max > 1
-    @property
-    def is_registration_on(self):   
-        return timezone.now() > self.registrarion_starts and timezone.now() < self.registrarion_ends
-        
-    # Other methods
-    def __unicode__(self):
-        return self.name
-
-class Tab(models.Model):
-    """
-        A model containing data for a tab that is displayed for the events on the mainsite
-    """
-    # Relation to various other models
-    event           = models.OneToOneField(Event, blank=True, null=True)
-    
-    # Baisc information
-    head            = models.CharField(max_length=100)
-    content         = models.TextField(blank = True)
-    
-    # Extra information about how the tab is shown on mainsite
-    order           = models.IntegerField(default=0, blank=False) # Tells which tab should come first in the list
-    is_visible      = models.BooleanField(default=True) # On the mainsite
-    
-    def __unicode__(self):
-        return self.title
-
-    class Meta:
-        ordering = ['order']
-
-
-
-#New
-
-
-class EventTab(models.Model):
-    """
-       Each event will have several tabs. This is one of them
-    """
-    # Relation to various other models
-    event           = models.ForeignKey(Event)
-    
-    # Baisc information
-    name            = models.CharField(max_length=100)
-    content         = models.TextField(blank = True)
-    
-
-    def __unicode__(self):
-        return self.name
-
-class EventRegistration(models.Model):
-    """
-        Each participant will have several fields 
-    """
-    event            = models.ForeignKey(Event, related_name='event_registered')
-    users_registered = models.ForeignKey(User, related_name='user_eventregis')
-    info             = models.TextField(null=True, blank=True)
-    timestamp        = models.DateTimeField(auto_now_add=True)
-    teams_registered = models.ForeignKey(Team, blank=True, null=True,related_name='user_team')
-    
-    def __unicode__(self):
-        return str(self.users_registered) +' - '+ str(self.event)
-
-class EventSchedule(models.Model):
-    """
-        each slot for an event will have a fields
-    """
-    event            = models.ForeignKey(Event, related_name='event_slot')
-    slot_start       = models.DateTimeField(null=True, blank=True)
-    slot_end         = models.DateTimeField(null=True, blank=True)
-    comment          = models.TextField(null=True, blank=True)
-    venue            = models.CharField(max_length=100, choices=EVENT_VENUES)
-    def __unicode__(self):
-        return str(self.event)
-
-
-class EventParticipation(models.Model):
-    event_for_participation = models.OneToOneField(Event, related_name='event_participated')
-    users_participated = models.ManyToManyField(UserProfile, related_name='user_participated',null=True, blank=True)
-    teams_participated = models.ManyToManyField(Team, related_name='team_participated',null=True, blank=True)
-    
-    
-    
-class EventWinner(models.Model):
-    """
-        each slot for a winner
-    """
-    event            = models.ForeignKey(Event, related_name='event_winners')
-    position         = models.CharField(null=True, blank=True, max_length=100)
-    comment          = models.CharField(null=True, blank=True,max_length=200)
-    added_by         = models.ForeignKey(ERPProfile,related_name='winners_added')
-    user             = models.ForeignKey(UserProfile, related_name='event_won')
-    def __unicode__(self):
-        return str(str(self.event)+'-'+str(self.user)+'-'+str(self.position))
-    
-
-@receiver(post_save, sender=EventSchedule)
-def generate_json(sender, **kwargs):
-    print "android post save signal"
-    call_command('android_json')
-    call_command('collectstatic', interactive=False)
-    if settings.PERMISSION_COMMAND:
-        os.system('/home/saarango/git/fest-api/runscript')
-
-@receiver(post_delete, sender=EventSchedule)
-def generate_json_delete(sender, **kwargs):
-    print "android postdelete signal"
-    call_command('android_json')
-    call_command('collectstatic', interactive=False)
-    if settings.PERMISSION_COMMAND:
-        os.system('/home/saarango/git/fest-api/runscript')
-class WebsiteUpdate(models.Model):
-    """
-        Update ticker on website
-    """
-    TYPE_CHOICES = (
-        ("info","info"),
-        ("success","success"),
-        ("warning","warning"),
-        ("error","error")
-    )
-    type = models.CharField(max_length=50,choices=TYPE_CHOICES)
-    title = models.CharField(max_length=100, blank=True, null=True)
-    text = models.TextField(max_length=2000, blank=True, null=True)
-
-    def __unicode__(self):
-        return self.title
-@receiver(post_save, sender=WebsiteUpdate)
-def generate_json_update(sender, **kwargs):
-    print "Updates post save signal"
-    call_command('updates_json')
-    call_command('collectstatic', interactive=False)
-    if settings.PERMISSION_COMMAND:
-        os.system('/home/saarango/git/fest-api/runscript')
-@receiver(post_delete, sender=WebsiteUpdate)
-def generate_json_update_delete(sender, **kwargs):
-    print "updates post delete signal"
-    call_command('updates_json')
-    call_command('collectstatic', interactive=False)
-    if settings.PERMISSION_COMMAND:
-        os.system('/home/saarango/git/fest-api/runscript')
