--- apps/hospi/forms.py
@@ -1,58 +1,0 @@
-from django import forms
-from apps.users.models import UserProfile
-from models import Hostel, Room, HospiTeam
-from misc.constants import *
-
-class HostelForm(forms.ModelForm):
-    class Meta:
-        model = Hostel
-
-class RoomForm(forms.ModelForm):
-    class Meta:
-        model = Room
-        exclude = ['occupants']
-
-class HospiTeamForm(forms.ModelForm):
-    class Meta:
-        model = HospiTeam
-        exclude = ['leader','team_sid', 'members', 'checked_in', 'checked_out', 'mattress_count', 'mattress_returned' ]
-    def __init__(self, *args, **kwargs):
-        super(HospiTeamForm, self).__init__(*args, **kwargs)
-
-
-class UserProfileForm(forms.Form):
-    first_name         = forms.CharField(max_length=30)
-    last_name          = forms.CharField(max_length=30)
-    email              = forms.EmailField(max_length=100)
-    age                = forms.CharField(max_length=2)
-    gender             = forms.ChoiceField(choices=GENDER_CHOICES)
-    mobile_number      = forms.CharField(max_length=15, help_text='Please enter your current mobile number')
-    branch             = forms.ChoiceField(choices=BRANCH_CHOICES, help_text='Your branch of study')
-    college_text       = forms.CharField(max_length=50)
-    college_roll       = forms.CharField(max_length=40)
-    '''
-    class Meta:
-        model = UserProfile
-        fields = [
-            'gender', 
-            'dob', 
-            'mobile_number',
-            'branch',
-            'college',
-            'college_roll',
-            'school_student',
-            'want_accomodation',
-        ]
-
-    def init(self, *args, **kwargs):
-        super(UserProfileForm, self).__init__(*args, **kwargs)
-        self.fields['dob'].input_formats = settings.DATE_INPUT_FORMATS
-
-    # def clean_dob(self):
-        
-    def save(self, commit=True, *args, **kwargs):
-        instance = super(UserProfileForm, self).save(commit=False, *args, **kwargs)
-        if commit:
-            instance.save()
-        return instance
-    '''
