+++ files/templates/portals/hospi/ohm.html
@@ -1,0 +1,39 @@
+{% load url from future %}
+<style type="text/css">
+.widemodal{
+  width: 1200px;
+  margin-left: -600px;
+}
+</style>
+<legend>OHM Details<div class="pull-right"></div></legend>
+<table id="ohmtable">
+  <thead>
+    <tr>
+      <th><span class="badge badge-info"><i class="icon-home icon-white"></i> I/O</span></th>
+      <th>ID</th>
+      <th>Name</th>
+      <th>Check in</th>
+      <th>Check out</th>
+      <th>No.</th>
+      <th>Days</th>
+      <th>OHM</th>
+    </tr>
+  </thead>
+  <tbody>
+    {% for i in data %}
+    {% if i.team.checked_out %}
+    <tr class="rowlink">
+      <td><span class="badge badge-{% if team.checked_in and not team.checked_out %}success{% elif team.checked_in and team.checked_out %}danger{% elif not team.checked_in and not team_checked_out %}inverse{% endif %}"><i class="icon-home icon-white"></i>{% if team.checked_in and not team.checked_out %} IN{% elif team.checked_in and team.checked_out %} OUT {% elif not team.checked_in and not team_checked_out %} ?{% endif %} </span></td>
+      <td style="color:black">{{ i.team.team_sid }}</td>
+      <td style="color:black">{{ i.team.name }}</td>
+      <td style="color:black">{{ i.team.date_of_arrival }} {{i.team.time_of_arrival}}</td>
+      <td style="color:black">{{ i.team.date_of_departure }} {{i.team.time_of_departure}}</td>
+      <td style="color:black">{{ i.team.get_total_count }}</td>
+      <td style="color:black">{{ i.dos }}</td>
+      <td style="color:black">{{ i.ohm }}</td>
+    </tr>
+    {% endif %}
+    {% endfor %}
+  </tbody>
+</table>
+<div id="team_details"></div>
